Add option for displaying seed name
